refactor classes and utils
